amend! Add Playwright-based UI tests
Add Playwright-based UI tests

Playwright (https://playwright.dev/) is a quite complete framework to
perform UI tests on websites.

We use this framework here to verify that:

- the search works

- switching between languages works

- following links from a translated manual page fall back to English if
  the linked-to manual page has no translation

- navigating between book sections colorizes the correct links in the
  drop-down

A new feature of the Hugo/Pagefind-based site is that the search
facility is language-aware. For this reason, we also verify that:

- search results are language-dependent

These tests can be run by first installing the `@playwright/test`
dependency, via `npm install`, and then invoking `npx playwright test`.

Helped-extensively-by: Max Schmitt <max@schmitt.mx>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>