amend! Transmogrify the Rake script to generate the Pro Git book sections
Transmogrify the Rake script to generate the Pro Git book sections

These book sections are not yet linked, as we still have to generate the
book's "front page".

Note: Since the contents of the book will be carried in this repository
as checked-in, pre-rendered HTML (and kept up to date by scheduled
GitHub workflows), they will have their home not in `content/` directly,
but instead in `external/book/`, using Hugo mounts. During development,
there was a brief debate whether to use submodules instead; In the end
the complexity was deemed too high a price by the person driving the
implementation, and a sparse checkout-friendly, single-repository
paradigm is used instead.

Based-on-feedback-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Well-needed-mediation-provided-by: Victoria Dye <vdye@github.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>