amend! Downloads (MacOS): show relative date
Downloads (MacOS): show relative date

This is yet another thing that was computed server-side but can be
computed on the client side instead.

The logic for the approximate date difference was inspired by
https://github.com/rails/rails/blob/4136145819ab/actionview/lib/action_view/helpers/date_helper.rb#L29-L158
(because this was used indirectly in the Rails app, via the
`time_ago_in_words` function).

However, rather than reimplementing all of that complicated logic, let's
implement a vastly simpler yet still similar logic that imitates how
that Rails function handles year-long differencess: up to 25% longer a
time difference and we say "about X ago", up to 25% shorter and we say
"almost X ago", otherwise we say "over X ago".

Co-authored-by: Matthias Aßhauer <mha1993@live.de>
Signed-off-by: Matthias Aßhauer <mha1993@live.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>